+++ plugins/processing/signal-processing/src/box-algorithms/basic/ovpCBoxAlgorithmIdentity.cpp
@@ -10,7 +10,7 @@ using namespace std;
 
 void CBoxAlgorithmIdentity::release() { delete this; }
 
-bool CBoxAlgorithmIdentity::processInput(const uint32_t /*index*/)
+bool CBoxAlgorithmIdentity::processInput(const size_t /*index*/)
 {
 	getBoxAlgorithmContext()->markAlgorithmAsReadyToProcess();
 	return true;
