+++ plugins/processing/signal-processing/src/box-algorithms/basic/ovpCBoxAlgorithmReferenceChannel.cpp
@@ -67,7 +67,7 @@ bool CBoxAlgorithmReferenceChannel::uninitialize()
 	return true;
 }
 
-bool CBoxAlgorithmReferenceChannel::processInput(const uint32_t /*index*/)
+bool CBoxAlgorithmReferenceChannel::processInput(const size_t /*index*/)
 {
 	this->getBoxAlgorithmContext()->markAlgorithmAsReadyToProcess();
 	return true;
