+++ plugins/processing/signal-processing/src/box-algorithms/filters/ovpCBoxAlgorithmSpatialFilter.cpp
@@ -173,7 +173,7 @@ bool CBoxAlgorithmSpatialFilter::uninitialize()
 	return true;
 }
 
-bool CBoxAlgorithmSpatialFilter::processInput(const uint32_t /*index*/)
+bool CBoxAlgorithmSpatialFilter::processInput(const size_t /*index*/)
 {
 	this->getBoxAlgorithmContext()->markAlgorithmAsReadyToProcess();
 	return true;
