+++ plugins/processing/signal-processing/src/box-algorithms/ovpCBoxAlgorithmSimpleDSP.h
@@ -19,7 +19,7 @@ namespace OpenViBEPlugins
 			void release() override { delete this; }
 			bool initialize() override;
 			bool uninitialize() override;
-			bool processInput(const uint32_t index) override;
+			bool processInput(const size_t index) override;
 			bool process() override;
 			void evaluate();
