+++ plugins/processing/signal-processing/src/box-algorithms/spectral-analysis/ovpCBoxAlgorithmSpectralAnalysis.cpp
@@ -88,7 +88,7 @@ bool CBoxAlgorithmSpectralAnalysis::uninitialize()
 	return true;
 }
 
-bool CBoxAlgorithmSpectralAnalysis::processInput(const uint32_t /*index*/)
+bool CBoxAlgorithmSpectralAnalysis::processInput(const size_t /*index*/)
 {
 	getBoxAlgorithmContext()->markAlgorithmAsReadyToProcess();
 	return true;
